Refactor hunting configuration and enhance manual control features
- Updated `config.json` to replace `allowed_maps` with `manual_control_maps`, allowing for better management of maps where automation is suspended.
- Introduced `aim_offsets` to specify vertical aim adjustments for certain mob names, improving targeting accuracy.
- Enhanced CLI arguments to support targeting all visible mobs or specific mob names, providing greater flexibility during hunts.
- Refactored `HuntController` and related policies to accommodate the new manual control logic, ensuring automation is paused on designated maps.
- Updated documentation to reflect changes in configuration options and their implications for bot behavior.

+++ src/ro_bot/core/projection/camera.py
@@ -38,7 +38,7 @@ class CameraProjection:
     def map_to_screen(
         self,
         player_cell: tuple[int, int],
-        mob_cell: tuple[int, int],
+        mob_cell: tuple[float, float],
         rect: WindowRect,
     ) -> tuple[int, int]:
         """Return absolute screen pixel for ``mob_cell`` under this projection."""
